// src/components/editor/types.ts
export interface PsdTemplate {
  id: string;
  name: string;
  thumbnailUrl: string;
  demoUrl: string;
  size: {
    width: number;
    height: number;
  };
  category: string;
  jsonPath?: string;
  canvasJson?: string | object;
  photographer?: string;
  photographerUrl?: string;
  pexelsUrl?: string;
}

export interface UserTemplate {
  id: string;
  name: string;
  thumbnailUrl: string;
  size: {
    width: number;
    height: number;
  };
  category: string;
  canvasJson: string;
  dateCreated: string;
}